requirements.txt
# Core app
streamlit
requests
python-dotenv
setuptools>=45
imageio-ffmpeg
audio_extract 
assemblyai
# Video/audio processing
moviepy
imageio-ffmpeg

# LLM (LangChain + OpenAI)
openai
langchain
import ffmpeg